requirements.txt
numpy==2.2.2
pandas==2.2.3
scikit_learn==1.6.1
scipy==1.15.1
torch==2.5.1
tqdm==4.61.1
xgboost==2.1.3
pyaml==20.4.0